fix: align finance academic request identifiers

+++ EduOS.Core/DTOs/Finance/FeeBillingDtos.cs
@@ -4,8 +4,8 @@ namespace EduOS.Core.DTOs.Finance;
 
 public sealed class SaveFeeStructureDto
 {
-    [Range(1, int.MaxValue)] public int AcademicYearId { get; set; }
-    [Range(1, int.MaxValue)] public int ClassId { get; set; }
+    [Range(1, long.MaxValue)] public long AcademicYearId { get; set; }
+    [Range(1, long.MaxValue)] public long ClassId { get; set; }
     [Range(1, long.MaxValue)] public long FeeHeadId { get; set; }
     [Range(typeof(decimal), "0", "1000000000")] public decimal Amount { get; set; }
 }
@@ -13,9 +13,9 @@ public sealed class SaveFeeStructureDto
 public sealed class GenerateStudentInvoicesDto
 {
     public Guid ClientRequestId { get; set; }
-    [Range(1, int.MaxValue)] public int AcademicYearId { get; set; }
-    [Range(1, int.MaxValue)] public int ClassId { get; set; }
-    [Range(1, int.MaxValue)] public int SectionId { get; set; }
+    [Range(1, long.MaxValue)] public long AcademicYearId { get; set; }
+    [Range(1, long.MaxValue)] public long ClassId { get; set; }
+    [Range(1, long.MaxValue)] public long SectionId { get; set; }
     [Range(1, 12)] public int Month { get; set; }
     [Range(2000, 2200)] public int Year { get; set; }
     public DateTime DueDate { get; set; }
